// src/modules/multichat/api/MultichatApiRoutes.ts
export enum MultichatApiRoutes {
  // chat providers
  GET_CHAT_PROVIDERS = '/api/chat/providers',
  // chat
  GET_MULTICHAT_UNSEEN_COUNT = '/api/chat/unseen-count',
  GET_CHAT = '/api/chat/chats/:chatId',
  DELETE_CHAT = '/api/chat/chats/:chatId',
  GET_CHATS = '/api/chat/chats',
  CREATE_PERSONAL_CHAT = '/api/chat/chats/personal',
  CREATE_GROUP_CHAT = '/api/chat/chats/group',
  UPDATE_GROUP_CHAT = '/api/chat/chats/group/:chatId',
  GET_CHAT_EXISTS = '/api/chat/chats/group/exists/:entityId',
  UPDATE_ALL_CHAT_MESSAGES_STATUS = '/api/chat/chats/:chatId/status/:status',
  FIND_ENTITY_CHATS = '/api/chat/chats/find',
  FIND_FULL_CHATS = '/api/chat/chats/find/full',
  FIND_FULL_CHATS_PERSONAL = '/api/chat/chats/find/full/personal',
  FIND_FULL_CHATS_BY_MESSAGE_CONTENT = '/api/chat/chats/find/full/by-message-content',
  CREATE_EXTERNAL_CHAT = '/api/chat/chats/external',
  // chat messages
  GET_CHAT_MESSAGES = '/api/chat/chats/:chatId/messages',
  GET_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages/:messageId',
  UPDATE_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages/:messageId',
  SEND_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages',
  DELETE_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages/:messageId',
  UPDATE_CHAT_MESSAGE_STATUS = '/api/chat/chats/:chatId/messages/:messageId/status/:status',
  UPDATE_CHAT_MESSAGES_STATUS = '/api/chat/chats/:chatId/messages/status/:status',
  REACT_TO_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages/:messageId/react/:reaction',
  UNREACT_TO_CHAT_MESSAGE = '/api/chat/chats/:chatId/messages/:messageId/unreact/:reactionId',
  // twilio
  GET_TWILIO_PROVIDERS_SETTINGS = '/api/chat/settings/providers/twilio',
  GET_TWILIO_PROVIDER_SETTINGS = '/api/chat/settings/providers/twilio/:providerId',
  UPDATE_TWILIO_PROVIDER_SETTINGS = '/api/chat/settings/providers/twilio/:providerId',
  CREATE_TWILIO_PROVIDER_SETTINGS = '/api/chat/settings/providers/twilio',
  DELETE_TWILIO_PROVIDER_SETTINGS = '/api/chat/settings/providers/twilio/:providerId',
  // facebook messenger
  GET_FB_MESSENGER_PROVIDERS_SETTINGS = '/api/chat/settings/providers/messenger',
  GET_FB_MESSENGER_PROVIDER_SETTINGS = '/api/chat/settings/providers/messenger/:providerId',
  GET_FB_MESSENGER_AUTH_REDIRECT_URL = '/api/chat/settings/providers/messenger/auth/connect',
  UPDATE_FB_MESSENGER_PROVIDER_SETTINGS = '/api/chat/settings/providers/messenger/:providerId',
  DELETE_FB_MESSENGER_PROVIDER_SETTINGS = '/api/chat/settings/providers/messenger/:providerId',
  // wazzup
  CREATE_WAZZUP_PROVIDER_SETTINGS = '/api/chat/settings/providers/wazzup',
  GET_WAZZUP_PROVIDERS_SETTINGS = '/api/chat/settings/providers/wazzup',
  GET_WAZZUP_PROVIDER_SETTINGS = '/api/chat/settings/providers/wazzup/:providerId',
  GET_WAZZUP_PROVIDER_SETTINGS_CHANNELS = '/api/chat/settings/providers/wazzup/channels',
  UPDATE_WAZZUP_PROVIDER_SETTINGS = '/api/chat/settings/providers/wazzup/:providerId',
  DELETE_WAZZUP_PROVIDER_SETTINGS = '/api/chat/settings/providers/wazzup/:providerId',
  GET_WAZZUP_API_KEY_BY_STATE = '/api/chat/settings/providers/wazzup/api-key',
  // create contact
  CREATE_CONTACT_FROM_CHAT = '/api/chat/chats/:chatId/contact',
}